move persistant iterator test
