+++ packages/native-core/Cargo.toml
@@ -26,3 +26,4 @@ dashmap = "5.4.0"
 
 [dev-dependencies]
 rand = "0.8.5"
+dioxus = { path = "../dioxus", version = "^0.2.1" }
